完善结构, 大部分功能完成

+++ brionac-backend/src/main/java/com/brionac/entity/domain/Specs.java
@@ -45,7 +45,7 @@ public class Specs implements Serializable {
      * 进价
      */
     @TableField(value = "specs_in_price")
-    private Double specsInPrice;
+    private BigDecimal specsInPrice;
 
     /**
      * 规格出售价格
